Fix sharing bug
Because we set the URL manually, thanks to Andrew's insistence on those awful `utm` query
strings, we need to update it manually as well. Added a way for the social module to expose
a refresh function, and removed the header share button just in case it got confused. Also
removed the top ad to keep marketing happy.